Update README.md
You have been hired by Six Flags to deliver the next big ride tracking system. The roller coasters and trolleys have been bought – your co-workers are just waiting for the code! And so, you set out to amusement parks to learn more about how you should implement the system.

Solution Description

You will create three classes: Ride.java, RollerCoaster.java, and Trolley.java. In this assignment, we use the following definitions:

A “stop” refers to an individual destination along the ride.

For roller coasters, a “stop” is synonymous with a “run”.

A “run” consists of stop(s) and refers to an entire loop on the ride.

For roller coasters, a run consists of one stop.

For trolleys, a run may consist of multiple stops.